don't depend on user's current date display format
If the current user's date display format is RFC2822 (e.g. "Mon, 15 Dec
2014 18:19:28 +1100"), pressing "+1d" may not correctly result in the
date advancing by 24 hours.

This seems to be a current limitation of the RT web interface's date
parser.

Enforcing the outputted new dates to be in ISO format (and thus not
relying on the user's display preference) seems to fix the problem.